HORRIBLE hack to order as intended.
Instead of being ordered by their parent resource, child resources, e.g.
'Creating a Service Object', are ordered by their URI. This leads to
ridiculous behavior, as the order in the API file doesn't match the
generated order.
To prevent this, you have to specify a URI for the parent resource
(which will be ignored) as well as for the child resource (which
actually will be used). This generates the intended order.